Se agregan manejo de excepciones y Se agrega logica de sanciones y logs con SLF4J para sanction-service

+++ sanction-service/sanction-service/src/main/java/cl/duoc/esports/sanctionservice/repositories/SancionRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface SancionRepository extends JpaRepository<Sancion,Integer> {
+public interface SancionRepository extends JpaRepository<Sancion, Long> {
 
     List<Sancion> findByUsuarioId(Long usuarioId);
